Various (rollup of stashes etc..) Including * Versioning and various Configuration * Added signing key * Added EfEagerLoadContext for carrying params across the query building core functions as well as for Filtering support * Added support for OnlyIfRootType * Redid Benchmarks - now in new application and now doing various comparisons * Started structure for the core library tests lib * Added commons library for Testing utilizies (not the tests themselves) * New Models for Testing library and ConsoleTesting setup code * Misc. code refactoring, restructuing and reformating * Add todo and removed "future wishes comment out of the "Production" code
