+++ docs/ToDo.md
@@ -1,0 +1,9 @@
+# EfEagerLoad 
+## (Still to evaluate and do)
+
+
+Add Unit Tests
+
+### Futures: 
+    // ToDO: Still need to include mechanism for having cached Func based on the ignoredNavigationProperties in addition to the type?
+    // (maybe even look into Predicate => Expression with matching)
