+++ source/Core/EfEagerLoad.Tests/Common/GuardTests.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace EfEagerLoad.Tests.Common
+{
+    class GuardTests
+    {
+    }
+}
