+++ source/Core/EfEagerLoad.Tests/EfQueryableExtensionsTests.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace EfEagerLoad.Tests
+{
+    public class EfQueryableExtensionsTests
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
