+++ source/Core/EfEagerLoad.Tests/Internal/EagerLoadAttributeFunctionsTests.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace EfEagerLoad.Tests.Internal
+{
+    class EagerLoadAttributeFunctionsTests
+    {
+    }
+}
