+++ source/Core/EfEagerLoad.Tests/Internal/InternalEfQueryableHelperTests.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace EfEagerLoad.Tests.Internal
+{
+    public class InternalEfQueryableHelperTests
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
