+++ source/Core/EfEagerLoad/Attributes/EagerLoadAttribute.cs
@@ -5,8 +5,14 @@ namespace EfEagerLoad.Attributes
     [AttributeUsage(AttributeTargets.Property)]
     public class EagerLoadAttribute : Attribute
     {
-        public EagerLoadAttribute()
+        public EagerLoadAttribute(bool onlyIfOnRoot = false, bool notIfOnRoot = false)
         {
+            OnlyIfOnRoot = onlyIfOnRoot;
+            NotIfOnRoot = notIfOnRoot;
         }
+
+        public bool OnlyIfOnRoot { get; }
+
+        public bool NotIfOnRoot { get; }
     }
 }
