+++ source/Core/EfEagerLoad/Common/Guard.cs
@@ -1,7 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System;
 
 namespace EfEagerLoad.Common
 {
