+++ source/Core/EfEagerLoad/Properties/PublishProfiles/FolderProfile.pubxml
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<!--
+https://go.microsoft.com/fwlink/?LinkID=208121. 
+-->
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <PublishProtocol>FileSystem</PublishProtocol>
+    <Configuration>Release</Configuration>
+    <Platform>Any CPU</Platform>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <PublishDir>bin\Release\netcoreapp3.1\publish\</PublishDir>
+  </PropertyGroup>
+</Project>
