+++ source/Testing/EfEagerLoad.Benchmarks/EfEagerLoad.Benchmarks.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="3.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\EfEagerLoad\EfEagerLoad.csproj" />
+    <ProjectReference Include="..\EfEagerLoad.Testing\EfEagerLoad.Testing.csproj" />
+  </ItemGroup>
+
+</Project>
