+++ source/Testing/EfEagerLoad.Benchmarks/Program.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using BenchmarkDotNet.Running;
+using EfEagerLoad.Benchmarks.Miscellaneous;
+
+namespace EfEagerLoad.Benchmarks
+{
+    public static class Program
+    {
+        static void Main(string[] args)
+        {
+            var summary = BenchmarkRunner.Run<General>();
+        }
+    }
+}
