+++ source/Testing/EfEagerLoad.ConsoleTester/EfEagerLoad.ConsoleTester.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="3.1.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\EfEagerLoad\EfEagerLoad.csproj" />
+    <ProjectReference Include="..\EfEagerLoad.Testing\EfEagerLoad.Testing.csproj" />
+  </ItemGroup>
+
+</Project>
