+++ source/Testing/EfEagerLoad.ConsoleTester/Logging/ConsoleLogger.cs
