+++ source/Testing/EfEagerLoad.ConsoleTester/Logging/ConsoleLoggerFactory.cs
