+++ source/Testing/EfEagerLoad.Testing/Model/Author.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using EfEagerLoad.Attributes;
+
+namespace EfEagerLoad.Testing.Model
+{
+    public class Author
+    {
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        [EagerLoad]
+        public virtual IList<Book> Books { get; set; } = new List<Book>();
+    }
+}
