+++ source/Testing/EfEagerLoad.Testing/Model/Book.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using EfEagerLoad.Attributes;
+
+namespace EfEagerLoad.Testing.Model
+{
+    public class Book
+    {
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        [EagerLoad]
+        public virtual Publisher Publisher { get; set; }
+
+        [EagerLoad]
+        public virtual Category Category { get; set; }
+
+        [EagerLoad]
+        public virtual Author Author { get; set; }
+    }
+}
