+++ source/Testing/EfEagerLoad.Testing/Repository/EntityFrameworkRepository.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+
+namespace EfEagerLoad.Testing.Repository
+{
+    public class EntityFrameworkRepository : IRepository
+    {
+        private readonly DbContext _dbContext;
+
+        public EntityFrameworkRepository(DbContext dbContext)
+        {
+            _dbContext = dbContext;
+        }
+
+        public async Task<IList<TEntity>> GetAll<TEntity>(bool eagerLoad = true, bool includeDeleted = false) where TEntity : class
+        {
+            return await _dbContext.Set<TEntity>().EagerLoad(_dbContext, eagerLoad).ToListAsync();
+        }
+
+        public async Task<IList<TEntity>> GetAllMatching<TEntity>(Expression<Func<TEntity, bool>> predicate, bool eagerLoad = true, bool includeDeleted = false) where TEntity : class
+        {
+            return await _dbContext.Set<TEntity>().Where(predicate).EagerLoad(_dbContext, eagerLoad).ToListAsync();
+        }
+
+        public IQueryable<TEntity> Query<TEntity>(bool eagerLoad = false, bool includeDeleted = false) where TEntity : class
+        {
+            return _dbContext.Set<TEntity>().EagerLoad(_dbContext, eagerLoad).AsQueryable();
+        }
+    }
+}
