--- source/Tests/EfEagerLoad.ConsoleTester/Data/TestDbContext.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using EfEagerLoad.ConsoleTester.Model;
-using Microsoft.EntityFrameworkCore;
-
-namespace EfEagerLoad.ConsoleTester.Data
-{
-    public class TestDbContext : DbContext
-    {
-
-        public TestDbContext(DbContextOptions<TestDbContext> options) : base(options)
-        {
-        }
-
-        public DbSet<Publisher> Publishers { get; set; }
-
-        public DbSet<Author> Authors { get; set; }
-
-        public DbSet<Book> Books { get; set; }
-
-        public DbSet<Category> Category { get; set; }
-
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-        }
-
-    }
-}
