--- source/Tests/EfEagerLoad.ConsoleTester/Model/Publisher.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Text;
-using EfEagerLoad.Attributes;
-
-namespace EfEagerLoad.ConsoleTester.Model
-{
-    public class Publisher
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        [EagerLoad]
-        public virtual IList<Book> Books { get; set; } = new List<Book>();
-    }
-}
