--- source/Tests/EfEagerLoad.ConsoleTester/Repository/IRepository.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-
-namespace EfEagerLoad.ConsoleTester.Repository
-{
-    public interface IRepository
-    {
-        Task<IList<TEntity>> GetAll<TEntity>(bool eagerLoad = true, bool includeDeleted = false) where TEntity : class;
-
-        Task<IList<TEntity>> GetAllMatching<TEntity>(Expression<Func<TEntity, bool>> predicate, bool eagerLoad = true,
-                                                        bool includeDeleted = false) where TEntity : class;
-
-        IQueryable<TEntity> Query<TEntity>(bool eagerLoad = true, bool includeDeleted = false) where TEntity : class;
-    }
-}
